Created a Scissors, Paper and Rock game
created a variable the holds the return value of a function called
computerGuess, this function contains an array containing paper,
scissors, and rock. And a random number generator called randomGenerator
that is passed as an index to the array called choice.

furthermore Two variable were created called computer and player and initialized to 0
 a function called playRound is created that accept two parameter
 playSelection and computerSelection, and a conditional that check which
 which is greater; if playSelection > computerSelection then 1 is added to
  player else 1 is added to computer.

finally a function called game is created that implement a loop five times that
and prompt a user to enter choice and playround function is called
is called  and a conditional statement that check the winner.